Separated classes into different files.
